Develop

+++ BackEnd/goorm/src/main/java/backend/goorm/diet/service/FoodService.java
@@ -1,20 +1,19 @@
 package backend.goorm.diet.service;
 
 import backend.goorm.diet.dto.FoodResponseDto;
+import backend.goorm.diet.dto.FoodUpdateRequestDto;
 import backend.goorm.diet.dto.FoodUserDto;
 import backend.goorm.diet.entity.Food;
 import backend.goorm.diet.repository.DietRepository;
 import backend.goorm.diet.repository.FoodRepository;
 import backend.goorm.member.model.entity.Member;
 import backend.goorm.member.repository.MemberRepository;
-import backend.goorm.s3.service.S3ImageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -26,58 +25,74 @@ public class FoodService {
     private final FoodRepository foodRepository;
     private final MemberRepository memberRepository;
     private final DietRepository dietRepository;
-    private final S3ImageService s3ImageService;
 
-    public List<FoodResponseDto> getFoodByName(Long id, String name) {
-        if (name == null) {
+    public List<FoodResponseDto> getFoodByName(Long memberId, String foodName) {
+        if (foodName == null) {
             return null;
         }
 
+
+        Member member = memberRepository.findByMemberId(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("Member not found with id: " + memberId));
+
+        log.info("memberId: {}", member);
+        log.info("foodName: {}", foodName);
         Pageable limit = PageRequest.of(0, 20);
-        List<Food> foods = foodRepository.findDistinctFoodNameByMember(name, null, limit);
+        List<Food> foods = foodRepository.findDistinctFoodNameByMember(foodName, member, limit);
         return FoodResponseDto.fromEntityList(foods);
     }
 
-    public List<FoodResponseDto> getRecentFood(Long id) {
+    public List<FoodResponseDto> getAllFoods() {
+        List<Food> foods = foodRepository.findAll();
+        return FoodResponseDto.fromEntityList(foods);
+    }
+
+    public List<FoodResponseDto> getRecentFood(Long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("Member not found with id: " + memberId));
+
         Pageable pageRequest = PageRequest.of(0, 20, Sort.Direction.DESC, "createdAt");
-        List<Food> foods = dietRepository.findDistinctFoodByMember(null, pageRequest);
+        List<Food> foods = dietRepository.findDistinctFoodByMember(member, pageRequest);
         return FoodResponseDto.fromEntityList(foods);
     }
 
-    public FoodResponseDto createFood(Long id, FoodUserDto dto ) {
+    public FoodResponseDto createFood(Long memberId, FoodUserDto dto) {
+        if (dto.getAmount() == null) {
+            throw new IllegalArgumentException("Amount cannot be null");
+        }
+
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("Member not found with id: " + memberId));
+
         Food food = dto.toEntity();
-
-//        if (image != null && !image.isEmpty()) {
-//            String imageUrl = s3ImageService.upload(image);
-//            food.setImageUrl(imageUrl);
-//        }
-
+        food.setMember(member);
         foodRepository.save(food);
         return FoodResponseDto.fromEntity(food);
     }
 
-    public FoodResponseDto updateFood(Long id, Long foodId, FoodUserDto dto) {
+    public FoodResponseDto updateFood(Long memberId, Long foodId, FoodUpdateRequestDto dto) {
+        if (dto.getAmount() == null) {
+            throw new IllegalArgumentException("Amount cannot be null");
+        }
+
         Food food = foodRepository.findById(foodId)
                 .orElseThrow(() -> new IllegalArgumentException("Food not found with id: " + foodId));
+
+        if (!food.getMember().getMemberId().equals(memberId)) {
+            throw new IllegalArgumentException("You do not have permission to update this food item.");
+        }
 
         dto.updateEntity(food);
-
-//        if (image != null && !image.isEmpty()) {
-//            String imageUrl = s3ImageService.upload(image);
-//            food.setImageUrl(imageUrl);
-//        }
-
         foodRepository.save(food);
         return FoodResponseDto.fromEntity(food);
     }
 
-    public boolean deleteFood(Long id, Long foodId) {
+    public boolean deleteFood(Long memberId, Long foodId) {
         Food food = foodRepository.findById(foodId)
                 .orElseThrow(() -> new IllegalArgumentException("Food not found with id: " + foodId));
 
-        String imageUrl = food.getImageUrl();
-        if (imageUrl != null && !imageUrl.isEmpty()) {
-            s3ImageService.deleteImageFromS3(imageUrl);
+        if (!food.getMember().getMemberId().equals(memberId)) {
+            throw new IllegalArgumentException("You do not have permission to delete this food item.");
         }
 
         foodRepository.delete(food);
